BARK-87: Move BarkReport into report-models
