+++ src/app/user/(logged-in)/my-pets/_lib/dog-viewer/dog-viewer-data.ts
@@ -1,4 +1,4 @@
-import { BarkReportSchema } from "@/lib/bark/models/bark-report";
+import { BarkReportSchema } from "@/lib/bark/models/report-models";
 import { DogAppointmentSchema } from "@/lib/bark/models/dog-appointment";
 import { DogPreferredVetSchema } from "@/lib/bark/models/dog-preferred-vet";
 import { DogProfileSpecSchema } from "@/lib/bark/models/dog-profile-models";
