+++ src/app/vet/_lib/components/report-card.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Badge } from "@/components/ui/badge";
-import { BarkReport } from "@/lib/bark/models/bark-report";
+import { BarkReport } from "@/lib/bark/models/report-models";
 import {
   REPORTED_INELIGIBILITY,
   ReportedIneligibility,
