+++ src/lib/bark/operations/op-fetch-report.ts
@@ -1,4 +1,4 @@
-import { BarkReport } from "../models/bark-report";
+import { BarkReport } from "../models/report-models";
 import { CODE } from "@/lib/utilities/bark-code";
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { BarkContext } from "@/lib/bark/bark-context";
