+++ src/lib/bark/operations/op-fetch-reports-by-vet-id.ts
@@ -1,6 +1,6 @@
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { BarkContext } from "../bark-context";
-import { BarkReport } from "../models/bark-report";
+import { BarkReport } from "../models/report-models";
 import { CODE } from "@/lib/utilities/bark-code";
 import { toBarkReport } from "../mappers/to-bark-report";
 import { selectReportsByVetId } from "../queries/select-reports-by-vet-id";
