+++ tests/bark/operations/op-fetch-report.test.ts
@@ -1,4 +1,4 @@
-import { BarkReport, BarkReportSchema } from "@/lib/bark/models/bark-report";
+import { BarkReport, BarkReportSchema } from "@/lib/bark/models/report-models";
 import { BarkReportData } from "@/lib/bark/models/bark-report-data";
 import { givenDog, givenReport, givenUser, givenVet } from "../_given";
 import { mockReportData } from "../_mocks";
